Lighthouse Computer Assisted Personal Interview
For those who use Lighthouse Computer Assisted Personal Interview system, the second file is a no-code solution for processing that difficult .txt output into an excel sheet. Since the only output from this outdated system is a difficult text, this AI .exe Applicaiton uses Natural Language Processing to extract the data from the output and organizes it in excel for easy upload to whatever database you may be using. As an app this is also a no code solution.